refactor handlers.py
