+++ vultron/api/v2/backend/handlers/case.py
@@ -1,0 +1,377 @@
+"""
+Handler functions for vulnerability case activities.
+"""
+
+import logging
+
+from vultron.api.v2.backend.handlers._base import verify_semantics
+from vultron.enums import MessageSemantics
+from vultron.types import DispatchActivity
+
+logger = logging.getLogger(__name__)
+
+
+@verify_semantics(MessageSemantics.CREATE_CASE)
+def create_case(dispatchable: DispatchActivity) -> None:
+    """
+    Process a CreateCase activity (Create(VulnerabilityCase)).
+
+    Persists the new VulnerabilityCase to the DataLayer, creates the
+    associated CaseActor (CM-02-001), and emits a CreateCase activity to
+    the actor outbox. Idempotent: re-processing an already-stored case
+    succeeds without side effects (ID-04-004).
+
+    Args:
+        dispatchable: DispatchActivity containing the as_Create with
+                      VulnerabilityCase object
+    """
+    from vultron.api.v2.data.rehydration import rehydrate
+    from vultron.api.v2.datalayer.tinydb_backend import get_datalayer
+    from vultron.behaviors.bridge import BTBridge
+    from vultron.behaviors.case.create_tree import create_create_case_tree
+
+    activity = dispatchable.payload
+
+    try:
+        actor = rehydrate(obj=activity.actor)
+        actor_id = actor.as_id
+        case = rehydrate(obj=activity.as_object)
+        case_id = case.as_id
+
+        logger.info("Actor '%s' creates case '%s'", actor_id, case_id)
+
+        dl = get_datalayer()
+        bridge = BTBridge(datalayer=dl)
+        tree = create_create_case_tree(case_obj=case, actor_id=actor_id)
+        result = bridge.execute_with_setup(
+            tree=tree, actor_id=actor_id, activity=activity
+        )
+
+        if result.status.name != "SUCCESS":
+            logger.warning(
+                "CreateCaseBT did not succeed for actor '%s' / case '%s': %s",
+                actor_id,
+                case_id,
+                result.feedback_message,
+            )
+
+    except Exception as e:
+        logger.error(
+            "Error in create_case for activity %s: %s",
+            activity.as_id,
+            str(e),
+        )
+
+
+@verify_semantics(MessageSemantics.ENGAGE_CASE)
+def engage_case(dispatchable: DispatchActivity) -> None:
+    """
+    Process an RmEngageCase activity (Join(VulnerabilityCase)).
+
+    The sending actor has decided to engage the case (RM → ACCEPTED). Records
+    their RM state transition in their CaseParticipant.participant_status.
+
+    RM is participant-specific: each CaseParticipant tracks its own RM state
+    independently of other participants in the same case.
+
+    Args:
+        dispatchable: DispatchActivity containing the as_Join with
+                      VulnerabilityCase object
+    """
+    from vultron.api.v2.data.rehydration import rehydrate
+    from vultron.api.v2.datalayer.tinydb_backend import get_datalayer
+    from vultron.behaviors.bridge import BTBridge
+    from vultron.behaviors.report.prioritize_tree import (
+        create_engage_case_tree,
+    )
+
+    activity = dispatchable.payload
+
+    try:
+        actor = rehydrate(obj=activity.actor)
+        actor_id = actor.as_id
+        case = rehydrate(obj=activity.as_object)
+        case_id = case.as_id
+
+        logger.info(
+            "Actor '%s' engages case '%s' (RM → ACCEPTED)", actor_id, case_id
+        )
+
+        dl = get_datalayer()
+        bridge = BTBridge(datalayer=dl)
+        tree = create_engage_case_tree(case_id=case_id, actor_id=actor_id)
+        result = bridge.execute_with_setup(
+            tree=tree, actor_id=actor_id, activity=activity
+        )
+
+        if result.status.name != "SUCCESS":
+            logger.warning(
+                "EngageCaseBT did not succeed for actor '%s' / case '%s': %s",
+                actor_id,
+                case_id,
+                result.feedback_message,
+            )
+
+    except Exception as e:
+        logger.error(
+            "Error in engage_case for activity %s: %s",
+            activity.as_id,
+            str(e),
+        )
+
+    return None
+
+
+@verify_semantics(MessageSemantics.DEFER_CASE)
+def defer_case(dispatchable: DispatchActivity) -> None:
+    """
+    Process an RmDeferCase activity (Ignore(VulnerabilityCase)).
+
+    The sending actor has decided to defer the case (RM → DEFERRED). Records
+    their RM state transition in their CaseParticipant.participant_status.
+
+    RM is participant-specific: each CaseParticipant tracks its own RM state
+    independently of other participants in the same case.
+
+    Args:
+        dispatchable: DispatchActivity containing the as_Ignore with
+                      VulnerabilityCase object
+    """
+    from vultron.api.v2.data.rehydration import rehydrate
+    from vultron.api.v2.datalayer.tinydb_backend import get_datalayer
+    from vultron.behaviors.bridge import BTBridge
+    from vultron.behaviors.report.prioritize_tree import create_defer_case_tree
+
+    activity = dispatchable.payload
+
+    try:
+        actor = rehydrate(obj=activity.actor)
+        actor_id = actor.as_id
+        case = rehydrate(obj=activity.as_object)
+        case_id = case.as_id
+
+        logger.info(
+            "Actor '%s' defers case '%s' (RM → DEFERRED)", actor_id, case_id
+        )
+
+        dl = get_datalayer()
+        bridge = BTBridge(datalayer=dl)
+        tree = create_defer_case_tree(case_id=case_id, actor_id=actor_id)
+        result = bridge.execute_with_setup(
+            tree=tree, actor_id=actor_id, activity=activity
+        )
+
+        if result.status.name != "SUCCESS":
+            logger.warning(
+                "DeferCaseBT did not succeed for actor '%s' / case '%s': %s",
+                actor_id,
+                case_id,
+                result.feedback_message,
+            )
+
+    except Exception as e:
+        logger.error(
+            "Error in defer_case for activity %s: %s",
+            activity.as_id,
+            str(e),
+        )
+
+    return None
+
+
+@verify_semantics(MessageSemantics.ADD_REPORT_TO_CASE)
+def add_report_to_case(dispatchable: DispatchActivity) -> None:
+    """
+    Process an AddReportToCase activity
+    (Add(VulnerabilityReport, target=VulnerabilityCase)).
+
+    Appends the report reference to the case's vulnerability_reports list
+    and persists the updated case to the DataLayer. Idempotent: re-adding a
+    report already in the case succeeds without side effects (ID-04-004).
+
+    Args:
+        dispatchable: DispatchActivity containing the as_Add with
+                      VulnerabilityReport object and VulnerabilityCase target
+    """
+    from vultron.api.v2.data.rehydration import rehydrate
+    from vultron.api.v2.datalayer.db_record import object_to_record
+    from vultron.api.v2.datalayer.tinydb_backend import get_datalayer
+
+    activity = dispatchable.payload
+
+    try:
+        report = rehydrate(obj=activity.as_object)
+        case = rehydrate(obj=activity.target)
+        report_id = report.as_id
+        case_id = case.as_id
+
+        dl = get_datalayer()
+
+        existing_report_ids = [
+            (r.as_id if hasattr(r, "as_id") else r)
+            for r in case.vulnerability_reports
+        ]
+        if report_id in existing_report_ids:
+            logger.info(
+                "Report '%s' already in case '%s' — skipping (idempotent)",
+                report_id,
+                case_id,
+            )
+            return None
+
+        case.vulnerability_reports.append(report_id)
+        dl.update(case_id, object_to_record(case))
+
+        logger.info("Added report '%s' to case '%s'", report_id, case_id)
+
+    except Exception as e:
+        logger.error(
+            "Error in add_report_to_case for activity %s: %s",
+            activity.as_id,
+            str(e),
+        )
+
+
+@verify_semantics(MessageSemantics.CLOSE_CASE)
+def close_case(dispatchable: DispatchActivity) -> None:
+    """
+    Process a CloseCase activity (Leave(VulnerabilityCase)).
+
+    Records that the sending actor is leaving/closing their participation
+    in the case. Emits an RmCloseCase activity to the actor outbox.
+
+    Args:
+        dispatchable: DispatchActivity containing the as_Leave with
+                      VulnerabilityCase object
+    """
+    from vultron.api.v2.data.rehydration import rehydrate
+    from vultron.api.v2.datalayer.db_record import object_to_record
+    from vultron.api.v2.datalayer.tinydb_backend import get_datalayer
+    from vultron.as_vocab.activities.case import RmCloseCase
+
+    activity = dispatchable.payload
+
+    try:
+        actor = rehydrate(obj=activity.actor)
+        actor_id = actor.as_id
+        case = rehydrate(obj=activity.as_object)
+        case_id = case.as_id
+
+        logger.info("Actor '%s' is closing case '%s'", actor_id, case_id)
+
+        dl = get_datalayer()
+
+        close_activity = RmCloseCase(
+            actor=actor_id,
+            object=case_id,
+        )
+        try:
+            dl.create(close_activity)
+            logger.info(
+                "Created RmCloseCase activity %s", close_activity.as_id
+            )
+        except ValueError:
+            logger.info(
+                "RmCloseCase activity for case '%s' already exists"
+                " — skipping (idempotent)",
+                case_id,
+            )
+            return None
+
+        actor_obj = dl.read(actor_id)
+        if actor_obj is not None and hasattr(actor_obj, "outbox"):
+            actor_obj.outbox.items.append(close_activity.as_id)
+            dl.update(actor_id, object_to_record(actor_obj))
+            logger.info(
+                "Added RmCloseCase activity %s to actor %s outbox",
+                close_activity.as_id,
+                actor_id,
+            )
+
+    except Exception as e:
+        logger.error(
+            "Error in close_case for activity %s: %s",
+            activity.as_id,
+            str(e),
+        )
+
+
+@verify_semantics(MessageSemantics.UPDATE_CASE)
+def update_case(dispatchable: DispatchActivity) -> None:
+    """
+    Process an UpdateCase activity (Update(VulnerabilityCase)).
+
+    Applies scalar field updates from the activity's object to the stored
+    VulnerabilityCase in the DataLayer. Restricted to the case owner: if
+    the sending actor is not the case owner, logs a WARNING and skips.
+    Idempotent: last-write-wins on scalar fields.
+
+    Args:
+        dispatchable: DispatchActivity containing the as_Update with
+                      VulnerabilityCase object
+    """
+    from vultron.api.v2.data.rehydration import rehydrate
+    from vultron.api.v2.datalayer.db_record import object_to_record
+    from vultron.api.v2.datalayer.tinydb_backend import get_datalayer
+    from vultron.as_vocab.objects.vulnerability_case import VulnerabilityCase
+
+    activity = dispatchable.payload
+
+    try:
+        actor_id = (
+            activity.actor.as_id
+            if hasattr(activity.actor, "as_id")
+            else str(activity.actor)
+        )
+        incoming = rehydrate(obj=activity.as_object)
+        case_id = (
+            incoming.as_id if hasattr(incoming, "as_id") else str(incoming)
+        )
+
+        dl = get_datalayer()
+        stored_case = dl.read(case_id)
+        if stored_case is None:
+            logger.warning(
+                "update_case: case '%s' not found in DataLayer — skipping",
+                case_id,
+            )
+            return None
+
+        owner_id = (
+            stored_case.attributed_to.as_id
+            if hasattr(stored_case.attributed_to, "as_id")
+            else (
+                str(stored_case.attributed_to)
+                if stored_case.attributed_to
+                else None
+            )
+        )
+        if owner_id != actor_id:
+            logger.warning(
+                "update_case: actor '%s' is not the owner of case '%s'"
+                " — skipping update",
+                actor_id,
+                case_id,
+            )
+            return None
+
+        if isinstance(incoming, VulnerabilityCase):
+            for field in ("name", "summary", "content"):
+                value = getattr(incoming, field, None)
+                if value is not None:
+                    setattr(stored_case, field, value)
+            dl.update(case_id, object_to_record(stored_case))
+            logger.info("Actor '%s' updated case '%s'", actor_id, case_id)
+        else:
+            logger.info(
+                "update_case: object for case '%s' is a reference only"
+                " — no fields to apply",
+                case_id,
+            )
+
+    except Exception as e:
+        logger.error(
+            "Error in update_case for activity %s: %s",
+            activity.as_id,
+            str(e),
+        )
